refactor(remotecontrol): split into Core/Server/UI assemblies, add Frame input-log foundation
Move HTTP server, handlers, protocol, and notification code into a new
Lilium.RemoteControl.Server assembly, separating transport concerns from
the domain layer (Core) and UI. Add a new Frame module (FrameClock,
FrameGate, InputFrame/Buffer/Sequencer/SymbolTable) laying the groundwork
for deterministic input recording and replay.

+++ Packages/RemoteControl/Core/Runtime/Utility/FrameRate.cs
@@ -32,9 +32,15 @@ namespace Lilium.RemoteControl
         /// <summary>
         /// FPS as a decimal value (e.g. 60.0 for 60fps).
         /// </summary>
+        /// <remarks>
+        /// The numerator is promoted to double before dividing. Dividing the two uints directly is
+        /// integer division, which truncated every rate that is not a whole number of frames per
+        /// second: 60000/1001 came back as 59 instead of 59.94, so a caller scaling a time by this
+        /// drifted by roughly one frame per second.
+        /// </remarks>
         public double AsDecimal()
         {
-            return denominator / numerator;
+            return denominator / (double)numerator;
         }
 
         /// <summary>
